fixed more warning messages of -Wsign-compare type
Import internal commit 15fa438ea7ed4fa099c8fe6d6b052c99df71ab0a

+++ groups/bsl/bslstl/bslstl_unorderedset.t.cpp
@@ -2498,7 +2498,7 @@ void TestDriver<KEY, HASH, EQUAL, ALLOC>::testCase22()
 
             ASSERTV(LINE, 0 == verifyContainer(Y, EXP, LENGTH));
             ASSERTV(LINE, da.numBlocksInUse(),
-                    2 * TYPE_ALLOC * LENGTH == da.numBlocksInUse());
+                    2 * TYPE_ALLOC * LENGTH == (size_t) da.numBlocksInUse());
 
             ObjStlAlloc mZ;  const ObjStlAlloc& Z = mZ;
 
@@ -2519,7 +2519,7 @@ void TestDriver<KEY, HASH, EQUAL, ALLOC>::testCase22()
             mX.insert(BEGIN, END);
             ASSERTV(LINE, 0 == verifyContainer(X, EXP, LENGTH));
             ASSERTV(LINE, da.numBlocksInUse(),
-                    TYPE_ALLOC * LENGTH == da.numBlocksInUse());
+                    TYPE_ALLOC * LENGTH == (size_t) da.numBlocksInUse());
         }
 
         CONT.resetIterators();
@@ -4994,7 +4994,7 @@ void TestDriver<KEY, HASH, EQUAL, ALLOC>::testCase7()
 
                 Obj Y11(X, &oa);
 
-                ASSERT(0 == LENGTH || oa.numBlocksTotal() > A);
+                ASSERT(0 == LENGTH || (size_t) oa.numBlocksTotal() > A);
 
                 // Due of pooling of memory alloctioon, we can't predict
                 // whether this insert will allocate or not.
